pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "apexflow"
version = "2.0.0a1"
description = "ApexFlow AI - Intelligent workflow automation"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "asyncpg>=0.29.0",
    "google-genai>=1.0.0",
    "google-cloud-aiplatform>=1.70.0",
    "networkx>=3.3",
    "apscheduler>=3.10.0",
    "httpx>=0.27.0",
    "firebase-admin>=6.5.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.5.0",
    "pyyaml>=6.0",
    "sse-starlette>=2.0.0",
    "python-multipart>=0.0.12",
    "python-dotenv>=1.0.0",
    "json-repair>=0.30.0",
    "numpy>=1.26.0",
    "Pillow>=10.0.0",
    "feedparser>=6.0.0",
    "beautifulsoup4>=4.12.0",
    "requests>=2.31.0",
    "aiohttp>=3.9.0",
    "readability-lxml>=0.8.0",
    "trafilatura>=1.6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.6.0",
    "mypy>=1.11.0",
    "pre-commit>=3.8.0",
    "alembic>=1.13.0",
    "psycopg2-binary>=2.9.9",
    "asyncpg-stubs>=0.29.0",
]

[tool.hatch.build.targets.wheel]
packages = ["core", "agents", "memory", "remme", "services", "tools", "routers", "shared", "config", "prompts"]

[tool.ruff]
target-version = "py312"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]

[tool.mypy]
python_version = "3.12"
strict = true
plugins = ["pydantic.mypy"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
